Add more refactor to remove clients directory, move it to Client & Converters
