+++ app/Keycloak/Listeners/CreateClients.php
@@ -7,6 +7,7 @@ namespace App\Keycloak\Listeners;
 use App\Domain\Integrations\Events\IntegrationCreated;
 use App\Domain\Integrations\Integration;
 use App\Domain\Integrations\Repositories\IntegrationRepository;
+use App\Keycloak\Client\ApiClient;
 use App\Keycloak\ClientCollection;
 use App\Keycloak\Config;
 use App\Keycloak\Events\MissingClientsDetected;
@@ -14,7 +15,6 @@ use App\Keycloak\Exception\KeyCloakApiFailed;
 use App\Keycloak\RealmCollection;
 use App\Keycloak\Repositories\KeycloakClientRepository;
 use App\Keycloak\ScopeConfig;
-use App\Keycloak\Service\ApiClient;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Psr\Log\LoggerInterface;
