+++ app/Keycloak/Listeners/UpdateClients.php
@@ -11,12 +11,12 @@ use App\Domain\Integrations\Events\IntegrationUrlUpdated;
 use App\Domain\Integrations\Integration;
 use App\Domain\Integrations\Repositories\IntegrationRepository;
 use App\Keycloak\Client;
+use App\Keycloak\Client\ApiClient;
 use App\Keycloak\Exception\KeyCloakApiFailed;
 use App\Keycloak\Repositories\KeycloakClientRepository;
 use App\Keycloak\ScopeConfig;
-use App\Keycloak\Service\ApiClient;
-use App\Keycloak\Service\IntegrationToKeycloakClientConverter;
-use App\Keycloak\Service\IntegrationUrlConverter;
+use App\Keycloak\Converters\IntegrationToKeycloakClientConverter;
+use App\Keycloak\Converters\IntegrationUrlConverter;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Psr\Log\LoggerInterface;
