+++ tests/Keycloak/Listeners/UpdateClientsTest.php
@@ -11,14 +11,14 @@ use App\Domain\Integrations\IntegrationUrl;
 use App\Domain\Integrations\IntegrationUrlType;
 use App\Domain\Integrations\Repositories\IntegrationRepository;
 use App\Keycloak\Client;
+use App\Keycloak\Client\ApiClient;
 use App\Keycloak\Config;
 use App\Keycloak\Listeners\UpdateClients;
 use App\Keycloak\RealmCollection;
 use App\Keycloak\Repositories\KeycloakClientRepository;
 use App\Keycloak\ScopeConfig;
-use App\Keycloak\Service\ApiClient;
-use App\Keycloak\Service\IntegrationToKeycloakClientConverter;
-use App\Keycloak\Service\IntegrationUrlConverter;
+use App\Keycloak\Converters\IntegrationToKeycloakClientConverter;
+use App\Keycloak\Converters\IntegrationUrlConverter;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
