+++ tests/Keycloak/Service/IntegrationUrlConverterTest.php
@@ -10,7 +10,7 @@ use App\Domain\Integrations\IntegrationUrl;
 use App\Domain\Integrations\IntegrationUrlType;
 use App\Keycloak\Client;
 use App\Keycloak\Realm;
-use App\Keycloak\Service\IntegrationUrlConverter;
+use App\Keycloak\Converters\IntegrationUrlConverter;
 use PHPUnit\Framework\TestCase;
 use Ramsey\Uuid\Uuid;
 use Ramsey\Uuid\UuidInterface;
