+++ tests/Keycloak/Service/KeycloakApiClientTest.php
@@ -7,12 +7,12 @@ namespace Tests\Keycloak\Service;
 use App\Domain\Integrations\Environment;
 use App\Domain\Integrations\Integration;
 use App\Keycloak\Client;
+use App\Keycloak\Client\KeycloakApiClient;
 use App\Keycloak\Config;
 use App\Keycloak\Exception\KeyCloakApiFailed;
 use App\Keycloak\Realm;
 use App\Keycloak\RealmCollection;
 use App\Keycloak\ScopeConfig;
-use App\Keycloak\Service\KeycloakApiClient;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\Psr7\Response;
 use PHPUnit\Framework\MockObject\MockObject;
